Treasure adjustments ready
* Made it so treasure adjustments are per treasure type

+++ EncounterGen/Selectors/Contracts/IRollSelector.cs
@@ -5,7 +5,7 @@ namespace EncounterGen.Selectors
     public interface IRollSelector
     {
         String SelectFrom(String baseRoll, Int32 modifier);
-        Int32 SelectFrom(String roll);
+        Double SelectFrom(String roll);
         String SelectRollFrom(String source);
     }
 }
